requirements.txt
streamlit
agno
groq
faiss-cpu
torch
transformers
pdfplumber
numpy
